requirements.txt
pandas>=1.4.1
numpy>=1.22.3
pytz>=2022.1
backports.zoneinfo>=0.2.1